# specs/003-resolve-critical-issues/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GASTAT International Dossier System API
  description: API for managing international relations, documents, and intelligence data
  version: 1.0.0
  contact:
    name: GASTAT Development Team
    email: dev@gastat.gov.sa
servers:
  - url: https://api.gastat.gov.sa/v1
    description: Production server
  - url: https://staging-api.gastat.gov.sa/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /events:
    get:
      summary: List events
      description: Get paginated list of events with filtering
      tags: [Events]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: event_type
          in: query
          schema:
            type: string
            enum: [meeting, conference, workshop, other]
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, completed, cancelled]
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create event
      description: Create a new event
      tags: [Events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '409':
          description: Event conflict detected
  /events/conflicts:
    get:
      summary: Check event conflicts
      description: Check for scheduling conflicts
      tags: [Events]
      parameters:
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: exclude_event_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conflict check results
          content:
            application/json:
              schema:
                type: object
                properties:
                  has_conflicts:
                    type: boolean
                  conflicts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'

  # Intelligence Reports Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        location:
          type: string
        event_type:
          type: string
          enum: [meeting, conference, workshop, other]
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        is_recurring:
          type: boolean
        recurrence_pattern:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateEventRequest:
      type: object
      required: [title, start_time, end_time, event_type]
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        location:
          type: string
        event_type:
          type: string
          enum: [meeting, conference, workshop, other]
        is_recurring:
          type: boolean
          default: false
        recurrence_pattern:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_prev:
          type: boolean
...